Add files via upload
r1 and r2 are two lines in the original image, each of which intersects with three lanes. This set is referred to as a "lane pixel set". It can be seen that in the same row, the distance(d1,d3) between lane pixel sets of different lanes is very long. And this distance is difficult to express in a regular way, for example d1 versus d2. Therefore, we use transformer because it can capture contextual information effectively